pyproject.toml
[project]
name = "costwatch"
version = "0.1.0"
description = "CostWatch — Serverless AWS Cost Dashboard"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "boto3>=1.34",
    "aws-cdk-lib>=2.140",
    "constructs>=10.0",
    "requests>=2.31",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "moto[dynamodb,ssm]>=5.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", "src"]
